+++ CHANGES.rst
@@ -11,6 +11,7 @@ Improvements
 ^^^^^^^^^^^^
 
 - Make list of event room bookings sortable (:issue:`4022`)
+- Log when a booking is split during editing (:issue:`4031`)
 
 Bugfixes
 ^^^^^^^^